Refactor RegexExpressionHelper
